Desktop/Project7/Project1.2/project/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
langgraph
langchain-huggingface
sentence-transformers
supabase
python-dotenv
pydantic
